+++ acme/tests/test_dask.py
@@ -51,6 +51,8 @@ def test_cluster_setup():
         cluster_cleanup()
         slurmOut = "/tmp/{}".format(getpass.getuser())
         client = slurm_cluster_setup(partition=defaultQ,
+                                     n_workers=1,
+                                     timeout=120,
                                      job_extra=["--output={}".format(slurmOut), "--job-name='averycustomjobname'"],
                                      interactive=False)
         assert "--output={}".format(slurmOut) in client.cluster.job_header
@@ -63,7 +65,11 @@ def test_cluster_setup():
             # Over-allocation of memory should default to partition max
             # (this should work on all clusters but we don't know partition
             # names, QoS rules etc.)
-            client = esi_cluster_setup(partition="8GBDEV", n_workers=1, mem_per_worker="9000MB", interactive=False)
+            client = esi_cluster_setup(partition="8GBDEV",
+                                       timeout=120,
+                                       n_workers=1,
+                                       mem_per_worker="9000MB",
+                                       interactive=False)
             memory = np.unique([w["memory_limit"] for w in client.cluster.scheduler_info["workers"].values()])
             assert memory.size == 1
             assert np.round(memory / 1000**3)[0] == 8
@@ -74,7 +80,11 @@ def test_cluster_setup():
             cluster_cleanup(client)
 
             # Specify CPU count manually
-            client = esi_cluster_setup(partition="16GBDEV", n_workers=1, n_cores=1, interactive=False)
+            client = esi_cluster_setup(partition="16GBDEV",
+                                       timeout=120,
+                                       n_workers=1,
+                                       n_cores=1,
+                                       interactive=False)
             assert [w["nthreads"] for w in client.cluster.scheduler_info["workers"].values()][0] == 1
             cluster_cleanup(client)
 
@@ -107,11 +117,12 @@ def test_backcompat_cluster():
 
     if useSLURM:
         client = slurm_cluster_setup(partition=defaultQ,
-                                    n_jobs=1,
-                                    workers_per_job=1,
-                                    mem_per_job="1GB",
-                                    n_jobs_startup=1,
-                                    interactive=False)
+                                     timeout=120,
+                                     n_jobs=1,
+                                     workers_per_job=1,
+                                     mem_per_job="1GB",
+                                     n_jobs_startup=1,
+                                     interactive=False)
         assert len(client.cluster.workers) == 1
         memory = [w["memory_limit"] for w in client.cluster.scheduler_info["workers"].values()]
         assert round(memory[0] / 1000**3) == 1